cdk/requirements.txt
aws-cdk-lib==2.26.0
aws-cdk.aws-lambda-python-alpha==2.26.0a0
pydicom==2.3.0